requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=3.0.0=pypi_0
blas=1.0=mkl
blinker=1.9.0=pypi_0
blosc=1.11.1=pypi_0
brotlicffi=1.0.9.2=py310h6a678d5_1
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.7.14=hbd8a1cb_0
catkin-pkg=1.0.0=pypi_0
certifi=2025.7.14=py310h06a4308_0
cffi=1.17.1=py310h1fdaa30_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.2.1=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
decorator=5.2.1=pypi_0
distro=1.9.0=pypi_0
docutils=0.21.2=pypi_0
exceptiongroup=1.3.0=pypi_0
executing=2.2.0=pypi_0
flask=3.1.1=pypi_0
fonttools=4.45.1=pypi_0
freetype=2.13.3=h4a9f257_0
future=1.0.0=py310h06a4308_0
gevent=25.5.1=pypi_0
greenlet=3.2.3=pypi_0
gunicorn=23.0.0=pypi_0
h5py=3.10.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.7=py310h06a4308_0
ikpy=3.4.2=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=8.37.0=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.5=pypi_0
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_3
libgcc-ng=15.1.0=h69a702a_3
libgomp=15.1.0=h767d61c_3
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx=15.1.0=h8f9b012_3
libstdcxx-ng=15.1.0=h4852527_3
libtiff=4.7.0=hde9077f_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.2=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
ninja=1.12.1=h06a4308_0
ninja-base=1.12.1=hdb19cb5_0
numpy=1.26.4=pypi_0
omegaconf=2.3.0=pypi_0
open-teach=1.0.0=dev_0
opencv-python=4.9.0.80=pypi_0
openjpeg=2.5.2=h0d4d230_1
openssl=3.5.1=h7b32b05_0
packaging=23.2=pypi_0
pandas=2.1.3=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=py310h06a4308_0
prompt-toolkit=3.0.51=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pybullet=3.2.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.19.2=pypi_0
pyparsing=3.1.1=pypi_0
pyrealsense2=2.54.2.5684=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch=1.12.1=cpu_py310hb1f1ab4_1
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.4=py310h06a4308_0
rospkg=1.6.0=pypi_0
scipy=1.11.4=pypi_0
setuptools=68.0.0=py310h06a4308_0
shapely=2.1.1=pypi_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.14.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchvision=0.13.1=cpu_py310h164cc8f_0
tqdm=4.66.1=pypi_0
traitlets=5.14.3=pypi_0
typing-extensions=4.12.2=py310h06a4308_0
typing_extensions=4.12.2=py310h06a4308_0
tzdata=2023.3=pypi_0
urllib3=2.5.0=py310h06a4308_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
zmq=0.0.0=pypi_0
zope-event=5.1=pypi_0
zope-interface=7.2=pypi_0
zstd=1.5.6=hc292b87_0
